Changed scores module to dictionary
