+++ frontend/src/store/modules/groups/actions.ts
@@ -4,7 +4,8 @@ import { RootState } from '@/store/types'
 import { Group } from '@/classes'
 
 // Default constant group until backend integration
-const defaultGroups = [new Group('The Unicorns', ['Yousef', 'Mariah', 'Daniel', 'Yael'], 'THE UNICORNS')]
+const defaultGroups = [new Group('The Unicorns', ['Yousef', 'Mariah', 'Daniel', 'Yael'], 'THE UNICORNS'),
+                       new Group('Awesomers', ['Yossi', 'Abed', 'Kobi', 'Ahmad'], 'AWESOMERS')]
 
 
 export const actions: ActionTree<GroupsState, RootState> = {
