+++ frontend/src/store/modules/scores/getters.ts
@@ -3,7 +3,12 @@ import { ScoresState } from './types'
 import { RootState } from '@/store/types'
 
 export const getters: GetterTree<ScoresState, RootState> = {
-  scores: (state: any) => {
+  scores: (state) => {
     return state.scores
+  },
+  solvedScores: (state) => (groupName: string) => {
+    return state.scores[groupName.toUpperCase()].filter((score) => {
+      return score.currentPoints === score.problem.points
+    })
   },
 }
