+++ frontend/src/store/modules/scores/types.ts
@@ -1,5 +1,6 @@
 import { Score } from '@/classes'
+import { Dictionary } from 'vuex'
 
 export interface ScoresState {
-    scores: Score[]
+    scores: Dictionary<Score[]>
 }
