+++ frontend/src/store/modules/user/getters.ts
@@ -6,4 +6,11 @@ export const getters: GetterTree<UserState, RootState> = {
     userGroup: (state, getters, rootState, rootGetters) => {
         return rootGetters['groups/group'](state.groupId)
     },
+    userScores: (state, getters, rootState, rootGetters) => {
+        if (state.groupId) {
+            return rootGetters.scores[state.groupId]
+        } else {
+            return []
+        }
+    },
 }
