+++ frontend/src/views/Home.vue
@@ -27,7 +27,7 @@
         </thead>
         <tbody>
           <!-- Fill with ScoreItems -->
-          <score-item v-for="(score, index) in $store.getters.scores" :key="score.problem.id" :score="score" :number="index + 1"></score-item>
+          <score-item v-for="(score, index) in $store.getters.userScores" :key="score.problem.id" :score="score" :number="index + 1"></score-item>
         </tbody>
       </table>
     </div>
